\name{SPSL-package}
\alias{SPSL-package}
\alias{SPSL}
\docType{package}
\title{ Site Percolation on Square Lattice (SPSL) }
\description{
SPSL package provides functionality for labeling of percolation cluster on 2D & 3D square lattice with various lattice size, relative fraction of accessible sites (occupation probability), iso- & anisotropy, von Neumann & Moore d-neighborhood.
}
\details{
\tabular{ll}{
Package: \tab SPSL\cr
Type: \tab Package\cr
Version: \tab 0.1-5\cr
Date: \tab 2012-05-14\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
\code{ssi20()} and \code{ssi30()} functions provides a sites labeling of the isotropic cluster on 2D & 3D square lattice with von Neumann 0-neighborhood.\cr
\code{ssa20()} and \code{ssa30()} functions provides a sites labeling of the anisotropic cluster on 2D & 3D square lattice with von Neumann 0-neighborhood.\cr
\code{ssi2d()} and \code{ssi3d()} functions provides a sites labeling of the isotropic cluster on 2D & 3D square lattice with Moore d-neighborhood.\cr
\code{ssa2d()} and \code{ssa3d()} functions provides a sites labeling of the anisotropic cluster on 2D & 3D square lattice with Moore d-neighborhood.\cr
}
\author{Pavel V. Moskalev <moskalefff@gmail.com>}
